feat(tags): enable tags and properties on non-markdown documents and email threads
